feat: Update job

+++ app/responses/job.py
@@ -24,8 +24,8 @@ class JobResponse(BaseResponse):
     company_id: int
     employment_type: int
     salary_type: Optional[int] = None
-    min_salary: Optional[float] = None
-    max_salary: Optional[float] = None
+    min_salary: Optional[int] = None
+    max_salary: Optional[int] = None
     currency_cd: Optional[str] = None
     city_id: Optional[int] = None
     location: Optional[str] = None
